pyproject.toml
[project]
name = "UoMarket"
version = "0.1.0"
description = "A trusted marketplace for UoM people"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=6.0",
    "loguru>=0.7.3",
    "pydantic>=2.12.5",
    "pyyaml>=6.0.3",
]
